refactor: extract formatSize to @repo/cli-utils
Move formatSize helper from document attachments command to the shared
cli-utils format module for reuse across commands.
