+++ packages/cli-utils/src/format.ts
@@ -1,3 +1,13 @@
 const formatDate = (dateStr: string): string => dateStr.slice(0, 10);
 
-export { formatDate };
+const formatSize = (bytes: number): string => {
+  if (bytes < 1024) {
+    return `${bytes} B`;
+  }
+  if (bytes < 1024 * 1024) {
+    return `${(bytes / 1024).toFixed(1)} KB`;
+  }
+  return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
+};
+
+export { formatDate, formatSize };
