+++ packages/cli-utils/src/index.ts
@@ -2,6 +2,6 @@ export { outputArgs, outputResult, pickFields, filterFields } from "./output";
 export { promptRequired, confirmOrExit } from "./prompt";
 export { readStdin } from "./stdin";
 export { splitArg } from "./split-arg";
-export { formatDate } from "./format";
+export { formatDate, formatSize } from "./format";
 export { printTable, type Column, type Row } from "./table";
 export { printDefinitionList, type DefinitionItem } from "./definition-list";
